+++ src/components/GoogleMap/index.ts
@@ -1,0 +1,6 @@
+export { GoogleMap } from "./GoogleMap";
+export { Directions } from "./Directions";
+export { MapHandler } from "./MapHandler";
+export { PlaceAutocomplete } from "./PlaceAutocomplete";
+export { CustomMapControl } from "./CustomMapControl";
+export { MarkerWithInfowindow } from "./MarkerWithInfowindow";
