--- src/components/GoogleMaps/PlaceSearch/PlaceSearch.module.css
@@ -1,10 +1,0 @@
-.searchInput {
-    margin-top: 10px;
-    width: 300px;
-    height: 40px;
-    font-size: 18px;
-}
-
-.map {
-    margin-top: 50px;
-}
