--- src/components/GoogleMaps/PlaceSearch/index.ts
@@ -1,1 +1,0 @@
-export { PlaceSearch } from "./PlaceSearch";
