--- src/components/GoogleMaps/TripRoute/index.ts
@@ -1,1 +1,0 @@
-export { TripRoute } from "./TripRoute";
