--- src/components/GoogleMaps/index.ts
@@ -1,1 +1,0 @@
-export { GoogleMaps } from "./GoogleMaps";
